requirements.txt
fastapi[standard]
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1